继续完善ios 64

+++ src/main/java/cn/banny/unidbg/file/AbstractFileIO.java
@@ -113,12 +113,12 @@ public abstract class AbstractFileIO implements FileIO {
     }
 
     @Override
-    public final int mmap2(Unicorn unicorn, long addr, int aligned, int prot, int offset, int length, Map<Long, MemoryMap> memoryMap) throws IOException {
+    public final long mmap2(Unicorn unicorn, long addr, int aligned, int prot, int offset, int length, Map<Long, MemoryMap> memoryMap) throws IOException {
         byte[] data = getMmapData(offset, length);
         unicorn.mem_map(addr, aligned, prot);
         memoryMap.put(addr, new MemoryMap(addr, aligned, prot));
         unicorn.mem_write(addr, data);
-        return (int) addr;
+        return addr;
     }
 
     protected byte[] getMmapData(int offset, int length) throws IOException {
