+++ src/main/java/cn/banny/unidbg/memory/Memory.java
@@ -25,7 +25,7 @@ public interface Memory extends IO, Loader {
 
     void setCallInitFunction();
 
-    int mmap2(long start, int length, int prot, int flags, int fd, int offset);
+    long mmap2(long start, int length, int prot, int flags, int fd, int offset);
     int mprotect(long address, int length, int prot);
     int brk(long address);
